[Refactor] MyNote - ViewController, MainView 분리
